Fixes frequency counting on normalized vs denormalized values
This change makes it so that the normalized and denormalized values of
identities on the daily identities pipeline works more similarly, by
filtering out the same values on both so that frequencies better match
between one and the other. This makes it so that calculating the more
frequent value later in the pipeline is more consistent between
normalized and denormalized values.